requirements.txt
# ── FastAPI 서버 ──────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6

# ── 설정 ─────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2

# ── DB 드라이버 ───────────────────────────────────────────
pyodbc==5.1.0          # SQL Server
psycopg2-binary==2.9.9 # PostgreSQL

# ── LLM ──────────────────────────────────────────────────
httpx==0.27.2          # HTTP 클라이언트 (Ollama)

# ── 유틸 ─────────────────────────────────────────────────
python-dotenv==1.0.1
